Cpu: Expand test coverage for addressing modes using LDA
LDA seems a like as good instruction as any to test most of these
addressing modes since it has a lot of them. I'll implement at least one
test for the other addressing modes as I come across them. I would like
to come up with a better way to setup these tests as all of the are
pretty much copy and paste (set up the Bus and Cpu, set the instruction
byte in memory, set the operand values in memory, etc.), but I'll roll
with it for now.