Cpu: Move address reading to individual instruction functions
I tried to offload the fetching of the instruction operands to the
Addressing Mode portion of the `execute` function, but that was causing
problems when trying to implement instructions like STA or JMP.

With this implementation, we can get the address and then just pass it
to the individual instruction functions. This allows them to do what
they want with the address like read from it or use for some other
purpose (like setting the PC for the JMP instruction).

One draw back to this implementation will be the .Accumulator addressing
mode. For instructions like ROL where it has Absolute and Zero Page
addressing modes AND an .Accumulator mode, the .Accumulator version will
probably have to be handled as a special case. I'm not thrilled about
that, but maybe a better way for handling it will show itself.