Cpu: Fix Indirect Address mode
We were just reading the first byte at the address without reading the
entire word. This patch corrects this by reading the entire word at the
effective address and setting it to `address`. As far as I can tell, it
looks like this Addressing Mode is only used by JMP.