Cpu: Add ADC Instruction
There may be some more edge cases that I can't think of, but we can
always add those later to the tests. ADC is tricky with all of the
status flags that can be set and the adding of three values.

Instead of using the Zig `@addWithOverflow` builtin to add two u8s, the
sum is stored in a u16 for two reasons.

Firstly, there are two opportunities for the sum to result in a carry:
once with the accumulator and the value being added and once when the
accumulator and carry are added together. This allows us to just sum all
three together and then check if we need to set the carry flag at the
end once.

Secondly, it looked confusing if we used the Zig builtin becuase, within
the context of the Zig builtin, 'overflow' refers to the bits of the
integer resulting in an overflow with a carry bit (u1), while 'overflow'
in the context of the program refers to the sign bit of the result of an
addition or subtraction not having the expected state (e.g. two negative
values added together resulting in a positive value).