Project: Add raylib
Added raylib to the 'libs' folder, added it as a git submodule to the
project, and included it as a link library in `build.zig`. This will be
needed when the PPU is implemented, but I just wanted to switch gears
and do something a little less monotonous than implementing the boring
execution instructions.