Cpu: Add helper methods to manage the stack
The stack on the Cpu will be managed through the `stackPush` and
`stackPop` functions. The stack pointer was originally 16 bits wide as
my original plan was to store it as an absolute address, but keeping it
as a 8 bit value is simpler and allows us to more easily keep track of
the stack size.