Cpu: Refactor .Relative addressing mode
The previous version was poorly written. It's been refactored so that
it's easier to read. Additionally, the old version had a subtle bug. If
you cast -128 from an i8 to a u8 by just multiplying -128 by -1, you'll
end up with 128 which is an invalid u8. So the old approach of making it
a positive u8 would have either caused a panic with the Zig safety
turned on or result in UB.

This version leverages the fact that the address is a u16. With this, we
can safely cast -128 * -1 into a u16 to correctly calculate the relative
address.