main: Add memory dump relative to the CPU's program counter
I wanted to have a way to step through the CPU and see where the program
counter's at along with the data around it. This commit adds a "window"
into memory so that I can see the surrounding bytes relative to the PC.
It's sort of reminiscent to Visual Studios memory debugging tools.