Cpu: Add or remove additional cycles from branching instructions
The only instructions that use the relative addressing mode are the
branching instructions. If branching occurs, then we add an additional
cycle. If the relative address crosses a page boundary, then another
additional cycle is added. However, if no branching occurs, then we
don't pay the page crossing penalty because the relative address is
never used.

This was implemented by passing a reference to `additional_cycles` to
all of the branching functions. This gives the branching instruction
functions the ability to directly modify `additional_cycles` based on
the branching condition.