main: Draw Cpu information to screen
This will be useful for debugging. All of the CPU registers, cycles, and
status flags are now displayed on the screen.

There was a lot of memory allocation going on everytime we wanted to
print something to the screen using `allocPrintZ` to format the string,
so a helper function named `drawText` was created to proxy a call to
raylib's `DrawText` function. It takes care of formating the string and
freeing the memory. Options that are needed for raylib's `DrawText` are
now handled through the `DrawTextOptions` struct.