Cartridge: Add struct to load a cartridge
This is a quick implementation just to get started. Lots of work still
needs to be done. Currently this is just a header and a block of memory.
The Cartridge doesn't really have 65535 bytes of data, but that was just
a large enough buffer to test loading data into it. The header of the
ROM is parsed and validated according to the .NES file format.

Eventually, the memory will be split into PRG ROM banks and CHR ROM
banks and correctly mapped to the Console.