main: Load a Cartridge and connect it to the console
The buffer had be increased since the program arguments can get quite
large. Here we'll try to read any arguments passed to the program and
treat them as filepaths to a ROM. Further error checking will to be
added like checking to make sure that the filepath exists and we can
report problems when creating the Cartridge such as invalid .NES ROM
signatures.