Console: Set system memory to 2kb in size
This breaks a lot of things. I probably shouldn't've set the memory on
the console to the full addressable 16 bit width to start with, but I
did it just to get started. In reality, the NES only has 2kb of real
memory while the 6kb after that are mirrors of the first 2kb, giving us
a total addressable RAM space of 8kb. This is why, in the `read` and
`write` functions, we check if the address is within the 8kb range, and
then bitwise AND the address by 0x07FF to ensure that it reads from the
2kb memory.

But this change broke a lot of things on the CPU. First, since the
`read` and `write` functions used to check if the address was in the
FULL u16 range (which it always will be of course... kind of a dumb
check in the first place), then it would read from memory. This meant
that, the function could NEVER error and the Zig optimizer would be like
"yeah, this always references a valid memory address, so it's void and
not !void" which meant we didn't have to handle the error when using
these functions. Well, now that we've removed that check and the
addressable RAM space is only 8kb, the optimizer can't make that
assumption that the address is always valid, and we have to handle the
invalid memory address errors everywhere we didn't. This required
cleaning up `Cpu.zig` to respect the potential errors that `read` and
`write` can throw. Additionally, some of the addresses used in the CPU
tests were not valid address ranges and they had to be changed. The BRK
instruction test has been removed for the time being due to the
cartridge being required to read the IRQ/BRK vetors.

Finally, the MemoryDebugDisplay was removed from `main.zig`. This was
cool, but Console.memory has changed so much that it wouldn't work. So
sad.