Cpu: Replace `init` with `reset`
When we want to reset the CPU and read from the reset vector, we'll now
call that from the `reset` function. This will perform a safety check to
ensure that a cartridge is available before reading from the reset
vector.

Since we have `reset`, we don't need to do anything fancy in an `init`
function, so it's been removed. Additionally, the reference that was
given to console via the `connectCpu` function was invalid anyways...