Console: Add reading/writing to cartridge PRG ROM banks
I'm not quite sure if this is right, but we'll roll with it for now. If
the PRG ROM size is 1, then our PRG ROM is only 16 KB, but the
addressable space is 32 KB. For Mapper 0 (which is what I'm trying to
imlpement), the data from the first 16 KB are mirrored giving us the
total range of 32 KB. So if the PRG ROM size is 1, and we're trying to
read from an address that's larger than 16 KB, then we need to read it's
mirror from the 16 KB that are available.

I'm not sure if I should modify Cartridge so that we encode the only two
possibilities for addressing the cartridge space. On the one hand, the
iNES spec says nothing about limiting the addressable space. It just
says that you multiple the PRG ROM size read from the header by 16 KB to
get the size of the cartridge; however, in practice, there are only 32
KB of addressable space... Hmmmm....