Cpu: Fix SBC and ADC executions
Running the nestest rom, SBC was implemented incorrectly. The carry flag
behaves as a "not borrowed" flag in SBC. Even going back to the SBC
instruction from my 6502 reference, it does say that it's the
accumualtor minus operand minus NOT carry.

https://www.masswerk.at/6502/6502_instruction_set.html#SBC

I also realized that I can figure out the overflow situation without
doing the calculation twice, so that was nice. I've cleaned that whole
situation up in SBC, so while I was at it, ADC was cleaned up to match
the cleaner SBC implementation.

Finally, since I learned that it's common to call SEC before calling
SBC, I've set the carry flag on the SBC tests so that the tests still
pass.