Cpu: Correct the RESET sequecne
I found a fantastic resource explaining what the 6502 goes through on
reset: https://www.pagetable.com/?p=410

In short, all interrupts on the 6502 go through a sequence of pushing
the PC and P register onto the stack. RESET does the same thing, but it
was "hacked" by setting the read/write line to "read" so that it goes
through the same motions as the other interrupts, without actually
pushing any values to the stack.

So RESET takes 7 cycles, the SP is decremented 3 times to seem like
there are three elements on the stack, and the interrupt disable flag is
set. This has now been added to `reset` on the Cpu.