NesLog: A wrapper around `nestest.log` to check CPU output
This is a debug tool to check the CPU's output against nestest.log. It's
been a really helpful tool for debugging the CPU so far. So why not
integrate it so that it's even MORE helpful?!? It'll print a debug
statement to the console when the CPU's state doesn't line up with
nestest.log's expected state.