Ppu: Add left and right pattern tables
Added left and right pattern tables to the PPU and a function to build
them. I'm not sure if this is the best way to store them yet since I can
imagine the pattern tables changing, but maybe it'll be okay (at least
for this mapper). For now I'll just build them once and display the
pattern tables to see what they look like.