Cpu: Make refactor ADC overflow check
Just made the check more explicit. We want to know if the 7th bit
(a.k.a. the signed bit) is set after seeing if the signed bit differs
between the sum & the A register and the sum & the operand. This is just
easier for me to read since I see that we're particularly looking if the
signed bit is set.