NesLog: Add nestest output test
When testing nestest.log, I was just making some changes in `main.zig`
and passing the path to the nestest rom when running the program. This
was starting to take a long time to run with all the rendering for each
frame with raylib.

This test was added to check the output the CPU with the nestest.log
more quickly since we can skip all of the rendering here. Since I'm
still working with nestest.log to make changes to the CPU, I changed the
test so that it reports what the difference is between the log and the
CPU before the test fails.