Cpu: Skip page crossing cycle penalty for write operation address modes
The 6502 has an optimization for certain address modes when the
operation is reading from memory. However, instructions using these
addressing modes that perform a write to memory skip this optimization
since writing to an incorrect address is bad. The cycle counts for the
instructions here already take into account the extra cycle because they
*must* read the full address and skip the optimization.