main: Draw NES display to a texture
Rather than drawing each pixel from the PPU buffer, it's way more
performant to set the pixels in a buffer, update a texture with the
buffer colors, and draw the texture once. From my benchmarks, this more
than doubled the FPS.

An added benefit to this approach is that we can continue to work with
the native NES display size (256 by 240) and just upscale the output by
enlarging the `output_display` rectangle if needed.