main: Have output_display x & y start at 0
I originally thought the x/y values here were used to set where the
Rectangle needs to be placed in the main window. But that's wrong.
That's set when drawing it in the window. It's used to set the initial
x/y values on the Rectangle (at least, that's what I think).