main: Use textures for left and right pattern tables
Drawing the pattern tables on every console step was *very* costly. This
is much more performant. Since the pattern tables are static for the
most part (at least with mapper 0 that's being implementing), the
pattern table textures only have to be built once when the program
starts. Now the two pattern tables can simply be drawn with one draw
call each on every frame.