Instruction: Add Remaining Illegal Opcodes
By having the instruction, address mode, and cycle count for the illegal
opcodes, we no longer have to handle the `InvalidInstruction` error.

All of these illegal opcodes are currently handled as a `NOP` on the
CPU. I may come back and fill in some of these illegal codes, but at
least the cycle count and instruction byte count reads are now correct.