Ppu: Fix `readPpudata` and `writePpudata`
These functions are ways for the CPU to interact with the PPU. Data is
read from an internal ppudata buffer which we now mimic. Each read
returns what was previously stored on the buffer and the following
`ppudata` value located at the `ppuaddr` is primed by being placed onto
the buffer.

Additionally, the `ppuaddr` is incremented after the `ppudata` register
is accessed. This increment changes based on the `ppuctrl` flag.