main: Use `DrawTexturePro` to draw the display output
The NES display should be larger than its 256x240 pixel display.
However, when rendering it, the 256x240 display was just repeated
across the `output_display` instead of being stretched.

Using `DrawTexturePro`, there is now an upscaled output display where
the NES display texture is upscaled to be twice as large. The output
display is now 245760 pixels instead of 61440 pixels repeated.