Ppu: Implement PPU rendering
Removed testing code that copied the nametable data into the display
buffer. Pixel data is now drawn to the display buffer as the PPU runs.
Rendering occurs here in the `step` function before the dots/scanlines
are incremented.

Some of the really basic PPU flag operations (such as the setting and
clearing of the vertical blank flag) have been added as well.