Ppu: Refactor `ptrn_tbl_idx` to `nametable_entry`
`ptrn_tbl_idx` isn't wrong since it is an index to an entry on the
pattern table, but all of the NES documentation simply refers to this
as the nametable entry.

Additionally, when reading the nametable entry, instead of reading
directly from the nametable array, it's been changed to use the `read`
function on the PPU so that we're reading from the nametable correctly
regardless if it's arranged vertically or horizontally.

Finally, more "compact" parts of this code have been broken up so it's
easier for me to understand what it's doing at a glance.