Ppu: Convert `ppuaddr` into a packed struct
This was done to easily address specific bits that make upthe PPUADDR.
Reading the NesDev wiki clarified a few things such as the `v` and `t`
registers really being alternative names for `ppuaddr` and a temporary
ppuaddr register.

I tried and tried to make the PpuAddress struct here be a union where I
could easily address the individual bits of the struct or address the
entire struct like in C/C++. I unforutnately could not do it as, with
safety on, the Zig runtime will panic when trying to access a different
active type of union. So it's just a packed struct, and everytime I want
to set all 16 bits on the struct, I have to use `@bitCast`.