Ppu: Read ppuaddr to locate nametable and attributes
Finding the nametable and attribute values from the dot + scanline was
good for initial testing, but it's time to read from the ppuaddr. It
gives us the nametable address to read from and the attribute byte can
be determined by looking at parts of the ppuaddr to find the attribute.

This breaks the current display as the ppuaddr isn't being correctly
incremented after the reads.