Ppu: Remove `framedata.attribute_entry`
This variable wasn't even being used. A new `u2` has been added to
`buffer` where we store the current attribute for the next tile. When
the attribute byte is fetched, we also calcluate which part of the byte
is needed for determining the color palette based on the quadrant. This
section has been refactored to be more clear.

Finally, the attribute LSB and MSB shifting has been simplified similar
to the BG pattern byte shifts.