main: Add BG and FG Palette viewer for debugging
While trying to fix a bug with the colors being rendered incorrectly, I
kept having this nagging feeling that the color palettes were being set
incorrectly. So to test this, I added a palette viewer so I can see if
it was true or not.

Turns out that the BG palette colors are being set correctly. At least I
can rule that possibility out and focus on other areas to debug this.