Ppu: Fix BG color palette fetching
The colors were off in the middle of the screen and it makes sense now.

When fetching the attribute byte, we shift the `coarse_y` bits to create
the offset into the attribute array at the bottom of the nametable. The
issue came from shifting the bits on `ppuaddr.coarse_y`. This variable
is a `u5`. We manipulate this by shifting the bits to the right by two
to give us the 3 MSBs on the variable. Then we shift the bits to the
left 3 so that these three bits occupy bit positions 3, 4, and 5.

But since the original value was a `u5`, shifting the `u5` left three
bits would push the MSB of the 3 bits that we needed off and we'd lose
this data. Thus, in our final address, the 5th bit was always a 0. I
incorrectly assumed that shifting was occurring on a `u16`, like the
final result was, but all of that shifting was happening on a `u5` which
resulted in the loss of the MSB.

The fix here was to cast `coarse_y` to a `u16` before doing any of the
shifting. For good measure, this same fix was applied to `coarse_x`,
even though we never shift those bits left.