Ppu: Correctly implement `oamdata` register
Previously, there was just a byte for storing `oamdata` before I fully
understood how this register worked. With the OAM storage now in place
on the PPU, the reading and writing to the `oamdata` register has been
correctly implemented.

Based on how the `Sprite` struct is laid out, there was a bit of work
needed to figure out which byte to read from and write to.