Console: Add cycle counter
Added a system clock to the Console itself keep the CPU and PPU in sync.
Previously, the CPU just had it's cycles and the Console would figure
out how many CPU cycles advanced to figure out how many PPU cycles
should occur since the PPU runs 3 times faster than the CPU. This by
itself was a bit awkward in the Console `step` function, but now this
approach is making the implementation of the DMA more complicated than
it should be. To make the DMA implementation easier, I was planning on
passing the number of cycles that occurred back from the CPU to the
Console, but it's just easier to leave this responsibility to the
Console. Moreover, the NES does have a system clock.

The old `cycles` variable on the CPU has been repurposed to
`cycles_remaining`. This allows the CPU to stay in sync with the system
clock. Since the Console keeps track of the cycles, the CPU may still be
in the middle of executing an instruction on any given tick. For every
Console tick, one PPU cycle will occur and a CPU cycle will occur every
three cycles.