main: Add timer to simulate the Console at 60 FPS
The timing has nearly always been off, but the timing problems became
more pronounced once the clock was added to the Console. The previous
method was very crude where we'd emulate the Console until the PPU frame
was "ready". This was done by seeing if the PPU was in the pre-render
scanline phase.

With this approach, we're emulating the Console for one second and
drawing the results. This gives us a rough 60 FPS which is close to the
NTSC timing. Using a delta time, we calculate how much time has elapsed
since the emulation started. After 1 second has passed, we render the
PPU buffer before continuing on. My guess is that, down the road, the
input will feel "off", and the code for handling controller input will
have to be accommodated alongside the Console emulation.