Ppu: Start rendering foreground sprites
To keep track of which sprites to render during the next scanline, the
`scanline_sprites` struct was added here. It has an array of indexes
into the OAM where we'll find the sprites to render during the next
scanline and a count.

During rendering, we spin through the foreground sprites to be rendered
during the scanline and overwrite the background tile. There is a lot
more work to be done, but at this point, sprites are being rendered at
the correct location on the screen -- although the colors and tiles are
way off at the moment.