Ppu: Fix horizontal sprite pixel reading
Renamed `diff` to `x_diff` since we'll need to start keeping track of
the `y_diff` during the sprite rendering.

The wrong pixel on the sprite was being accessed because the `x_diff`
wasn't being used to offset the hi and lo pixel values. This was fixed
by adding a `pixel_offset` and using that to find the correct bit on
`pixel_lo` and `pixel_hi`.