Ppu: Add `y_diff` to sprite rendering
Uses the `y_diff` between the scanline and sprite.y to work out the row
of bytes needed to make up the pixels. Prior to this, we were just
grabbing the first 2 bytes that made up the sprite. With this, we will
move onto the next row of sprites depending on the `y_diff` to correctly
render the sprite (at least, more correctly).

Still a work in progress as we'll need to check and see if the sprite is
flipped vertically.