Ppu: Render BG pixel if FG pixel color is transparent
The BG and FG pixel are now calcluated at the same time. We start by
assuming that both pixels are transparent. If the FG pixel is
transparent (0x00), then we'll fall back to using the BG pixel that was
calcluated and write it to the buffer once at the end of both
calcluations. There are more priorities to consider, but this is a
decent start.