Ppu: Render BG sprite if FG sprite pixel is transparent
Consolidated variables used to determine the pixel color. There is now
one `pixel` variable that's used to determine the pixel color rather
than trying to figure out which pixel "wins" at the end of the
evalution.

The BG pixel is assumed to be drawn by default. If there is a sprite at
this location, then it'll be shown based on its priority. If the FG
pixel is transparent, then the BG pixel will be shown since that's
what's behind the sprite.