Project: Upgrade project to use Zig 0.13.0
Added a `build.zig.zon` file so the project's compatible with Zig
0.13.0. `build.zig` also had some breaking changes since std.Build has
changed since 0.11.0. Part of these changes involved changing how raylib
is added to the build in conjunction with `build.zig.zon`.

The only breaking change in the code with this upgrade is that the Zig
compiler enforces variables to be `const` if they are not changed, so
that change was made across the project.

raylib had to be updated as well so that the build script would be
compatible with 0.13.0. Finally, Zig 0.13.0 outputs the cache now into a
hidden .zig-cache directory, so the .gitignore file was updated to
include that change.