Ppu: Use the backdrop value as the default pixel value
If the background pixel is transparent, then we need to fallthrough and
use the backdrop pixel value (which is palette 0, index 0). If the
bg_pixel is itself 0, then the bg_pixel value is "transparent" and we
need to fall back to using the backdrop pixel value. We'll start the
pixel evaluation by assuming the the pixel is this backdrop pixel value.