Ppu: Fix invalid sprite rendering
The PPU could get into a state where it would try and render a sprite at
the bottom of the screen during the rendering of the first row of
pixels. This occurred because the sprite evaluation wasn't happening
during the pre-render scanline. This caused old sprites from the bottom
of the screen of the last frame to be part of the sprite rendering of
the new frame.

The fix was to move the sprite evaluation out of the "visible scanlines"
section and into a place where it'd run when the scanline is -1. The
long chained `if/else if` statement was broken up into standalone `if`
statements so that more than one branch can be ran like in the case of
the pre-render scanline.