Ppu: Set the nametable bits on the temp ppuaddr on writes to PPUCTRL
There was a bug where the base nametable address was always 0x2000. This
was because, when the nametable address is written to the PPUCTRL
register, the nametable address wasn't also being set on the PPUADDR.

This is why scrolling between two nametables would work just fine, but
the background would snap back to what's at address 0x2000 once the X
scroll exceeded 255 on horizontal arrangement.