Ppu: Fix which cycles are used to draw the frame
The motivation for this came from two bugs. One bug was with one pixel
column "glitches" on the left and right side of the frame. The second
bug was with sprites being incorrectly rendered one pixel to the left.

The problem was that dot/cycle 0 was being drawn as part of the frame
despite dot 0 being an idle cycle and the final dot of 256 was not being
drawn. Once the visible dots were sorted and dot 0 was compensated for
during sprite rendering, the bug with the sprites being incorrectly
rendered one pixel too early was fixed.