main: Move output texture and buffer variables into a Display struct
The motivation behind this is to wrangle all the variables used to
render a frame in one place. The goal will be to move the debug display
variables into this struct which can be controlled by the debug flag.

The `output_img` has now always been unloaded as soon as the output
texture has been created as that is no needed once the texture has been
created. No need to keep it around in memory until the end of the
program.