Ppu: Use a double buffer to prevent screen tearing
The screen tearing occurred due to the PPU writing data for the next
frame while the previous frame was being rendered. The double buffer
fixes the problem by having one buffer that's designated as the "reader"
and one that's designated as a "writer". The "reader" will always hold a
complete frame while the "writer" can be continuously written to until
the frame is complete. Once the "writer" buffer is finished, the
"reader" and "writer" buffers are swapped to begin the next frame.