main: Remove NesLog from main binary
NesLog is only for testing. There were still remnants of it in `main`
that really don't belong in the binary. NesLog is imported in the `test`
block in main so it's included in the test suite.

The build script has been modified accordingly to exclude nestest.log
from the main binary but include it as part of the unit test build.