NesLog: Use Console to compare the state
The `cycle` variable no longer exists on the CPU, so we need to use the
Console to compare the cycle counter with nestest.log cycles. Since
Console runs three times faster than the CPU, the CPU cycle count is
calculated using the Console. The `compare` function now accepts a
reference to the `Console` instead of a `Cpu`.

Small fix ups such swapping the `connect[X]` functions with reference
assignments were also included here so that NesLog actually works. One
advantage to using this approach is that we no longer need to discard
the first call to `neslog.next()` as a special case. The clocking of the
Console/CPU happens right along side with the fetching of next line of
the nestest.log.