Cpu: Always decrement `cycles_remaining` count in `step`
By not decrementing `cycles_remaining` in the fetch/decode/execute
branch of `step`, the CPU timing was off by an additional cycle for
every instruction that was initially processed due to the
fetch/decode/execute branch was not being counted as a cycle. This meant
that, over time, the CPU timing would fall out of sync with the PPU
until it lapped the PPU and it would momentarily be in sync again. This
error was not noticed until the the NesLog test suite was fixed.

Fixing this has fixed some more subtle bugs. One bug was that the PPU
would have noticeable lag spikes that would occur at seemingly random
times. This was due to the PPU being heavily out of sync with the CPU
until the CPU would "lap" the PPU and it would be in sync again.
Additionally, since the CPU would be running faster than it should be,
the sprite hit zero detection would be picked up too late by CPU thereby
failing to set the nametable at the correct time. This led to the PPU
reading from the incorrect nametable after sprite hit zero detection for
many scanlines causing glitchy artifacts on the screen. Finally, since
the CPU would gradually be one cycle ahead, the emulation would feel too
fast and unnatural.

The CPU tests have been fixed here. Tests that were not explicitly
testing the cycle count have been removed. Tests that were testing cycle
counts have been modified by decrementing the expected cycle count by
one to accommodate the CPU step itself as a cycle.