main: Move debug pattern table building from PPU
This function is used for debugging purposes only and it doesn't really
belong on the PPU. Since we're only dealing with mapper 0, it's easy
just to build these debugging pattern tables once at the start of the
program if debug mode is enabled.

Since moving it over, the helper `setupPatternTables` function has been
removed since we can just call the `buildPatternTable` function twice in
this one place. Moreover, we now pass in the slice of CHR_ROM that's
relevant for the left/right pattern tables as opposed to a memory offset
into the CHR_ROM.